+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "antnest-chatbot",
-  "version": "1.0.7",
+  "version": "1.0.8",
   "description": "",
   "main": "index.js",
   "scripts": {
